20241102 우승훈 Error Handler Refactoring
=====Back-End=====
@ControllerAdvice Class 는 common.exception package 에 GlobalExceptionsHandler.java 하나로 통합.
둘 이상 컨트롤러에서 발생할 가능성이 있는 Exception 은 상기 클래스에서 Handle.
이 외 커스텀 예외는 발생하는 컨트롤러에서 @ExceptionHandler Annotation Method 로 처리.
