+++ package.json
@@ -4,7 +4,7 @@
   "main": "index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-"build-css" : "sass --watch src/main/resources/static/scss:src/main/resources/static/css/screen"
+    "build-css" : "sass --watch src/main/resources/static/scss:src/main/resources/static/css/screen"
   },
   "keywords": [],
   "author": "",
