+++ src/main/java/com/guncat/ecommerce/security/config/SecurityConfig.java
@@ -1,6 +1,5 @@
 package com.guncat.ecommerce.security.config;
 
-import com.guncat.ecommerce.security.handler.CustomAuthenticationFailureHandler;
 import com.guncat.ecommerce.users.enums.Role;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -11,6 +10,7 @@ import org.springframework.security.config.Customizer;
 import org.springframework.security.config.annotation.authentication.configuration.AuthenticationConfiguration;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
+import org.springframework.security.config.annotation.web.configurers.AbstractHttpConfigurer;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
@@ -25,7 +25,6 @@ import org.springframework.security.web.authentication.AuthenticationFailureHand
 public class SecurityConfig {
 
     private final UserDetailsService userDetailsService;
-    private final AuthenticationFailureHandler authenticationFailureHandler;
 
     @Bean
     public SecurityFilterChain securityFilterChain(HttpSecurity http) throws Exception {
@@ -43,8 +42,7 @@ public class SecurityConfig {
                         .loginPage("/users/login").permitAll()
                         .loginProcessingUrl("/users/login").permitAll()
                         .usernameParameter("userId").passwordParameter("password")
-                        .defaultSuccessUrl("/", true)
-                        .failureHandler(authenticationFailureHandler))
+                        .defaultSuccessUrl("/", true))
                 .rememberMe( remember -> remember
                         .rememberMeParameter("rememberMe")
                         .tokenValiditySeconds(7 * 24 * 60 * 60)
