--- src/main/java/com/guncat/ecommerce/security/handler/CustomAuthenticationFailureHandler.java
@@ -1,36 +1,0 @@
-package com.guncat.ecommerce.security.handler;
-
-import com.guncat.ecommerce.security.enums.CustomAuthenticationException;
-import com.guncat.ecommerce.users.exception.UserNotFoundException;
-import jakarta.servlet.ServletException;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.security.core.AuthenticationException;
-import org.springframework.security.web.authentication.AuthenticationFailureHandler;
-import org.springframework.security.web.authentication.SimpleUrlAuthenticationFailureHandler;
-import org.springframework.stereotype.Component;
-import org.springframework.web.bind.annotation.ControllerAdvice;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.ResponseBody;
-
-import java.io.IOException;
-import java.io.PrintWriter;
-
-@Component
-public class CustomAuthenticationFailureHandler implements AuthenticationFailureHandler {
-
-    @Override
-    public void onAuthenticationFailure(HttpServletRequest request,
-                                        HttpServletResponse response,
-                                        AuthenticationException e) throws IOException, ServletException {
-        CustomAuthenticationException c = CustomAuthenticationException.findException(e);
-
-        response.setStatus(c.getStatusCode());
-        PrintWriter writer = response.getWriter();
-        writer.print(c.getMessage());
-        writer.flush();
-    }
-
-}
